+++ KruacentExiled/KE.Items/Items/Molotov.cs
@@ -10,11 +10,12 @@ using Player = Exiled.API.Features.Player;
 using MEC;
 using UnityEngine;
 using PlayerRoles;
+using KE.Items.ItemEffects;
 
 namespace KE.Items.Items
 {
     [CustomItem(ItemType.GrenadeFlash)]
-    public class Molotov : CustomGrenade, ILumosItem
+    public class Molotov : CustomGrenade, ILumosItem, ICustomItem
     {
         public override uint Id { get; set; } = 1049;
         public override string Name { get; set; } = "Cocktail Molotov";
@@ -24,8 +25,7 @@ namespace KE.Items.Items
         public override bool ExplodeOnCollision { get; set; } = true;
         public float DamageModifier { get; set; } = 0f;
         public UnityEngine.Color Color { get; set; } = UnityEngine.Color.yellow;
-        private const float RefreshRate = 0.5f;
-        private const float Duration = 20f;
+        public CustomItemEffect Effect { get; set; }
         public override SpawnProperties SpawnProperties { get; set; } = new SpawnProperties()
         {
             Limit = 2,
@@ -69,88 +69,16 @@ namespace KE.Items.Items
             },
         };
 
+
+        public Molotov()
+        {
+            Effect = new MolotovEffect();
+        }
+
         protected override void OnExploding(ExplodingGrenadeEventArgs ev)
         {
-            float cylinderSize = 5;
-
-            ev.TargetsToAffect.Clear();
-
-            Player playerThrowingGrenade = ev.Player;
-            Vector3 molotovPosition = ev.Position;
-            Primitive wall = Primitive.Create(PrimitiveType.Cylinder, molotovPosition, null, new Vector3(cylinderSize, 0.01f, cylinderSize), true);
-            wall.Collidable = false;
-            wall.Visible = true;
-
-            wall.Color = Color.red;
-
-            var coroutineHandler = Timing.RunCoroutine(DamageInMolotovZone(wall.Position, cylinderSize, playerThrowingGrenade));
-
-            Timing.CallDelayed(Duration, () => {
-                wall.UnSpawn();
-                Timing.KillCoroutines(coroutineHandler);
-                wall.Destroy();
-            });          
+            //Effect.OnExploding(ev);
         }
 
-        private IEnumerator<float> DamageInMolotovZone(Vector3 wallPosition, float cylinderSize, Player playerThrowingGrenade)
-        {
-            // Dictionary that stores the time each player has spent inside the zone (in seconds).
-            Dictionary<Player, float> playerTimeInZone = new Dictionary<Player, float>();
-
-            while (true)
-            {
-                foreach (Player player in Player.List)
-                {
-                    if (IsPlayerInZone(player, wallPosition, cylinderSize))
-                    {
-                        if (Exiled.API.Features.Server.FriendlyFire || playerThrowingGrenade.Role.Team != player.Role.Team || playerThrowingGrenade == player)
-                        {
-                            if (player.IsHuman || player.Role == RoleTypeId.Scp0492)
-                            {
-                                if (playerTimeInZone.ContainsKey(player))
-                                {
-                                    // increase time each frame.
-                                    playerTimeInZone[player] += Time.deltaTime; 
-                                }
-                                else
-                                {
-                                    // Init the time in dictionnary of the player.
-                                    playerTimeInZone[player] = Time.deltaTime; 
-                                }
-
-                                // time of player spend inside of molotov zone.
-                                float timeInZone = playerTimeInZone[player];
-
-                                // Beginning it willbe 5dm/s, after it will be linearly higher the damage until 20dm/s. 
-                                float damage = Mathf.Lerp(2.5f, 10f, timeInZone / 20f);
-
-                                // double damage if it's zombie cuz it has more hp.
-                                if (player.Role == RoleTypeId.Scp0492)
-                                {
-                                    damage *= 2.5f;
-                                }
-
-                                player.Hurt(damage, DamageType.Bleeding);
-                            }
-                            else if (player.IsScp)
-                            {
-                                player.Hurt(player.Health / 150, DamageType.Bleeding);
-                            }
-
-                        }
-                    }
-                }
-
-                yield return Timing.WaitForSeconds(RefreshRate);
-            }
-        }
-
-
-        private bool IsPlayerInZone(Player player, Vector3 zonePosition, float radius)
-        {
-            float distance = Vector3.Distance(new Vector3(player.Position.x, 0, player.Position.z),
-                                               new Vector3(zonePosition.x, 0, zonePosition.z));
-            return distance <= (radius/2) ;
-        }
     }
 }
